Refactor helper functions into own util files
